+++ ChallengeWithUs/.settings/org.eclipse.core.resources.prefs
@@ -1,2 +1,3 @@
 eclipse.preferences.version=1
 encoding//src/main/webapp/WEB-INF/views/yj/MemberRegForm.jsp=UTF-8
+encoding//src/main/webapp/WEB-INF/views/yj/returnMemberId.jsp=UTF-8
