+++ ChallengeWithUs/src/main/java/com/smart/home/controller/ChallengesController.java
@@ -28,19 +28,21 @@ public class ChallengesController {
 	ChallengesService service;
 	
 	// 챌린지 리스트 조회
-	@GetMapping("/ChallengesList")
-	public ModelAndView ChallengesList(ChallengesPagingDTO pDTO) {
+//	@GetMapping("/ChallengesList")
+//	public ModelAndView ChallengesList(ChallengesPagingDTO pDTO) {
+//
+//		pDTO.setTotalRecord(service.ChallengesTotalRecord(pDTO));
+//		List<ChallengesDTO> list = service.ChallengesList(pDTO);
+//		
+//		ModelAndView mav = new ModelAndView();
+//
+//		mav.addObject("list", list);
+//		mav.addObject("pDTO", pDTO);
+//		mav.setViewName("yj/ChallengesList");
+//		return mav;
+//	}
+	
 
-		pDTO.setTotalRecord(service.ChallengesTotalRecord(pDTO));
-		List<ChallengesDTO> list = service.ChallengesList(pDTO);
-		
-		ModelAndView mav = new ModelAndView();
-
-		mav.addObject("list", list);
-		mav.addObject("pDTO", pDTO);
-		mav.setViewName("yj/ChallengesList");
-		return mav;
-	}
 	
 	// 챌린지 더보기
 	@GetMapping("/ChallengesListMore")
