+++ ChallengeWithUs/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -33,8 +33,8 @@
 		<beans:property name="host" value="smtp.naver.com" />
 		<beans:property name="port" value="587" />
 		<beans:property name="username"
-			value="" />
-		<beans:property name="password" value="" />
+			value="winterer601@naver.com" />
+		<beans:property name="password" value="dbswo601!19" />
 		<beans:property name="defaultEncoding" value="UTF-8" />
 		<beans:property name="javaMailProperties">
 			<beans:props>
