+++ ChallengeWithUs/src/main/webapp/WEB-INF/views/yj/returnMemberId.jsp
@@ -1,14 +1,5 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-</head>
+
 <body>
 	<c:if test="${memberId != null}">${memberId}</c:if>
 	<c:if test="${memberId == null}">아이디 없음</c:if>
 </body>
-</html>
