+++ ChallengeWithUs/src/main/webapp/inc/top.jspf
@@ -35,4 +35,12 @@
         </div>
         <a href="/home"><img class="logo" alt="Logo" src="<%= request.getContextPath() %>/imgs/logo.png" /></a>
     </div>
+    <c:if test="${logId==null }">
+    로그인중아님
+    </c:if>
+        <c:if test="${logId!=null }">
+                <div class="div-wrapper">
+                <div class="element"><a href="/home/register/logout">로그아웃</a></div>
+            </div>
+    </c:if>
 </nav>
